+++ src/pages/Instagram.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import Igsearch from "../components/Igsearch";
+function Instagram() {
+
+        return(
+            <div>
+                <Igsearch />
+            </div>
+        )
+
+}
+
+export default Instagram;