+++ src/pages/index.js
@@ -1,5 +1,6 @@
 export { default as Home } from './Home';
 export { default as Search } from './Search';
 export { default as Contact } from './Contact';
-export { default as Stats } from './Stats'
+export { default as Stats } from './Stats';
+export { default as Instagram } from './Instagram';
 export { default as Error } from './Error';